Refs #3966 : almost done with UI
need to add some scrollbar and some header and to handle the update action

+++ meveo-api/src/main/java/org/meveo/api/git/GitRepositoryApi.java
@@ -141,7 +141,9 @@ public class GitRepositoryApi extends BaseCrudApi<GitRepository, GitRepositoryDt
 
     public void remove(String code) throws BusinessException {
         final GitRepository repository = gitRepositoryService.findByCode(code);
-        gitRepositoryService.remove(repository);
+        if(repository != null) {
+            gitRepositoryService.remove(repository);
+        }
     }
 
     @Override
@@ -188,6 +190,10 @@ public class GitRepositoryApi extends BaseCrudApi<GitRepository, GitRepositoryDt
     @Override
     public GitRepositoryDto find(String code) throws MeveoApiException, EntityDoesNotExistsException {
         final GitRepository repository = code.equals(meveoRepository.getCode()) ? meveoRepository : gitRepositoryService.findByCode(code);
+        if(repository == null) {
+        	throw new EntityDoesNotExistsException(GitRepository.class, code);
+        }
+        
         return toDto(repository);
     }
